+++ lib/implementations/checkdigit/checkdigit.factory.d.ts
@@ -1,4 +1,3 @@
-import Hashids from 'hashids';
 import { OidFactory } from '../oid-factory.interface';
 import { Oid } from '../../oid';
 export declare class CheckdigitOidFactory implements OidFactory {
@@ -17,5 +16,5 @@ export declare class CheckdigitOidFactory implements OidFactory {
         scope: string;
         id: string | number;
     };
-    getEncoder(scopename: string): Hashids;
+    getEncoder(scopename: string): any;
 }