+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+* `PlaidIdentity` oid; shortcode: `pident`
+* `PlaidIdentityEntry` oid; shortcode: `pidentry`
+### Removed
+### Changed
+### Fixed
+  * launch.json boilerplate for new debugger
+### Deprecated
+### Security
+
+
 ## [4.0.5] -- 2020-06-09
 
 ### Added